Feature/tela splash

+++ lib/data/repositories/profile_repository/profile_repository_imp.dart
@@ -1,18 +1,25 @@
 import 'package:firebase_auth/firebase_auth.dart';
 
 import '../../../domain/model/user_model.dart';
+import '../../../ui/core/enum/preference_keys_enum.dart';
 import '../../model/user_response_model.dart';
 import '../../services/api_client/api_client.dart';
+import '../../services/shared_prefs_helper/shared_preference_helper.dart';
 import 'profile_repository.dart';
 
 class ProfileRepositoryImp implements ProfileRepository {
   final ApiClient _apiClient;
   final FirebaseAuth _auth;
 
-  ProfileRepositoryImp(
-      {required ApiClient apiClient, required FirebaseAuth auth})
-      : _apiClient = apiClient,
-        _auth = auth;
+  final SharedPreferenceHelper _sharedPreferenceHelper;
+
+  ProfileRepositoryImp({
+    required ApiClient apiClient,
+    required FirebaseAuth auth,
+    required SharedPreferenceHelper sharedPreferenceHelper,
+  })  : _apiClient = apiClient,
+        _auth = auth,
+        _sharedPreferenceHelper = sharedPreferenceHelper;
 
   @override
   Future<UserModel> getUser() async {
@@ -33,24 +40,51 @@ class ProfileRepositoryImp implements ProfileRepository {
   @override
   Future<void> editProfile(String name) async {
     await _apiClient.patch('/users/updateMe', body: {
-      'data': {'name': name},
+      'data': {'username': name},
     });
   }
 
   @override
-  Future<void> userDelete(String userId) async {
-    await _auth.currentUser?.delete();
-    await _apiClient.delete('/users/$userId');
+  Future<void> userDelete(int? userId) async {
+    if (userId == null) {
+      throw Exception('User not authenticated');
+    }
+    try {
+      final email =
+          _sharedPreferenceHelper.getString(PreferenceKeysEnum.userEmail.name);
+      final password = _sharedPreferenceHelper
+          .getString(PreferenceKeysEnum.userPassword.name);
+
+      if (email != null && password != null) {
+        await _auth.signInWithEmailAndPassword(
+          email: email,
+          password: password,
+        );
+      }
+      await _apiClient.delete('/users/$userId');
+      await _auth.currentUser?.delete();
+    } catch (e) {
+      throw Exception('User not authenticated');
+    }
   }
 
   @override
-  Future<void> changePassword(String password) async {
+  Future<void> changePassword(
+      {required String password, required String newPassword}) async {
+    final email =
+        _sharedPreferenceHelper.getString(PreferenceKeysEnum.userEmail.name);
+    if (email != null) {
+      await _auth.signInWithEmailAndPassword(
+        email: email,
+        password: password,
+      );
+    }
     final user = _auth.currentUser;
 
     if (user != null) {
-      await user.updatePassword(password);
+      await user.updatePassword(newPassword);
       await _apiClient.patch('/users/updateMe', body: {
-        'data': {'password': password},
+        'data': {'password': newPassword},
       });
     } else {
       throw Exception('User not authenticated');
